requirements.txt
pytest==8.0.2
python-dotenv==1.0.1
requests==2.31.0
selenium==4.18.1
